Split files with multiple @parallel tests
